requirements.txt
fastapi
uvicorn
pydantic>=2
python-dotenv
scikit-learn
numpy
pandas
tqdm
openai
